% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classificationPlot.R
\name{classificationPlot}
\alias{classificationPlot}
\title{Plot a bidimensional matrix}
\usage{
classificationPlot(kohonenObject, cluster, matchingColor = "blue",
  nonMatchingColor = "white", ...)
}
\arguments{
\item{kohonenObject}{the kohonen class object to plot}

\item{cluster}{the cluster where the object has been put.}

\item{matchingColor}{the colors to be used for the matching cluster. Default is blue.}

\item{nonMatchingColor}{the colors to be used for non matching clusters. Default is white.}

\item{...}{other arguments to pass to the lower functions}
}
\value{
nothing: just perform the plot
}
\description{
Visualize where an object has been placed by kohonen's SOM.
}
\seealso{
plot.kohonen
}

