% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootCauseBarPlot.R
\name{rootCauseBarPlot}
\alias{rootCauseBarPlot}
\title{Bar plot}
\usage{
rootCauseBarPlot(x, y, title, xlab = "x", ylab = "y")
}
\arguments{
\item{x}{values to be plotted on the x axes}

\item{y}{values to be plotted on the y axes}

\item{title}{the title of the plot}

\item{xlab}{labels of the x axes}

\item{ylab}{labels of the y axes}
}
\value{
nothing, just perform the plot
}
\description{
Performs a bar plot
}

